rewrote backtesting with Luno API decimals package
