+++ go/src/backtesting-manuj/indicators.go
@@ -1,0 +1,11 @@
+package main
+
+/*Indicators that we can reuse. E.g. SMA*/
+
+func SMA(array []decimal.Decimal) decimal.Decimal{
+  sum := decimal.Zero()
+  for _, val := range array {
+    sum.Add(val)
+  }
+  return sum.Div(len(array))
+}
