--- go/src/backtesting/main.go
@@ -1,20 +1,0 @@
-package main
-
-import ()
-
-var priceCol int
-var timeCol int
-var sheetNum int
-
-
-func main () {
-
-  // initialising global variables
-  priceCol = 7
-  timeCol = 0
-  sheetNum = 0
-
-//	var getNextPrice ticker = tickerFunc()
-
-	tester()
-}
